#69 move classes into core
