+++ self-core-impl/src/test/java/com/selfxdsd/core/managers/StoredProjectManagerTestCase.java
@@ -20,12 +20,14 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.selfxdsd.api.storage;
+package com.selfxdsd.core.managers;
 
 import com.selfxdsd.api.Project;
 import com.selfxdsd.api.ProjectManager;
 import com.selfxdsd.api.Projects;
 import com.selfxdsd.api.Repo;
+import com.selfxdsd.api.storage.Storage;
+import com.selfxdsd.core.projects.StoredProject;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
